add remote login status and login prompt Fixes #54
