+++ src/tree_data_providers/conan_remote_item.ts
@@ -7,9 +7,9 @@ export class ConanRemoteItem extends vscode.TreeItem {
     ) {
         super(remote.name, vscode.TreeItemCollapsibleState.None);
         
-        this.contextValue = 'remote';
+        this.contextValue = remote.requires_auth ? 'remote-login-required' : 'remote';
         this.iconPath = new vscode.ThemeIcon('globe');
         this.tooltip = `Remote: ${remote.name}\nURL: ${remote.url}`;
-        this.description = remote.url;
+        this.description = remote.requires_auth ? '(requires authentication)' : '';
     }
 }
